chore(zero-cache): rename pg/tables package to pg/schema
