+++ packages/zero-cache/src/services/change-streamer/pg/initial-sync.test.ts
@@ -17,7 +17,7 @@ import type {
 import {Database} from 'zqlite/src/db.js';
 import {initialSync, replicationSlot} from './initial-sync.js';
 import {fromLexiVersion} from './lsn.js';
-import {getPublicationInfo} from './tables/published.js';
+import {getPublicationInfo} from './schema/published.js';
 
 const REPLICA_ID = 'initial_sync_test_id';
